refactor: make options validation as separate module
