+++ ciphering-cli-tool/lib/validate/index.js
@@ -1,0 +1,15 @@
+const {
+  validateConfig,
+  validateInput,
+  validateOutput,
+} = require('./validate-helpers')
+
+const { ValidationError } = require('./validate-errors')
+
+module.exports = {
+  validateConfig,
+  validateInput,
+  validateOutput,
+  
+  ValidationError,
+}
