+++ ciphering-cli-tool/lib/validate/validate-errors.js
